Rough Draft Complete

+++ controllers/categoryCtrl.js
@@ -1,4 +1,5 @@
 const Category = require('../models/categoryModel');
+const Products = require('../models/productModel');
 
 const categoryCtrl = {
   getCategories: async (req, res) => {
@@ -11,7 +12,8 @@ const categoryCtrl = {
   },
   createCategory: async (req, res) => {
     try {
-      // Admin Rights (role=1) - Create, Delete, Update Category
+      // if user have role = 1 ---> admin
+      // only admin can create , delete and update category
       const { name } = req.body;
       const category = await Category.findOne({ name });
       if (category)
@@ -20,15 +22,21 @@ const categoryCtrl = {
       const newCategory = new Category({ name });
 
       await newCategory.save();
-      res.json({ msg: 'Created category' });
+      res.json({ msg: 'Created a category' });
     } catch (err) {
       return res.status(500).json({ msg: err.message });
     }
   },
   deleteCategory: async (req, res) => {
     try {
+      const products = await Products.findOne({ category: req.params.id });
+      if (products)
+        return res.status(400).json({
+          msg: 'Please delete all products with a relationship.',
+        });
+
       await Category.findByIdAndDelete(req.params.id);
-      res.json({ msg: 'Deleted Category' });
+      res.json({ msg: 'Deleted a Category' });
     } catch (err) {
       return res.status(500).json({ msg: err.message });
     }
@@ -38,7 +46,7 @@ const categoryCtrl = {
       const { name } = req.body;
       await Category.findOneAndUpdate({ _id: req.params.id }, { name });
 
-      res.json({ msg: 'Updated Category' });
+      res.json({ msg: 'Updated a category' });
     } catch (err) {
       return res.status(500).json({ msg: err.message });
     }
